feat: add treatment when try to delete a nonexists university

+++ src/modules/universities/services/DeleteUniversityByIdService.ts
@@ -1,6 +1,7 @@
 import { getManager } from 'typeorm';
 import { ObjectID } from 'mongodb';
 import { University } from '../typeorm/entities/University';
+import AppError from '@shared/errors/AppError';
 
 interface IRequest {
   id: string;
@@ -10,7 +11,14 @@ export class DeleteUniversityByIdService {
   public async execute({ id }: IRequest): Promise<any> {
     const manager = getManager();
 
+    const universityExists = await manager.findOne(University, {
+      _id: new ObjectID(id),
+    });
+    if (!universityExists) {
+      throw new AppError('Resource not found', 404);
+    }
+
     await manager.delete(University, { _id: new ObjectID(id) });
-    return 
+    return;
   }
 }
